requirements.txt
ccxt>=2.0.0
python-dotenv>=0.19.0
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
streamlit>=1.10.0  # Optional for dashboard
loguru>=0.5.3  # Enhanced logging
psutil>=5.9.0  # System monitoring
requests>=2.27.0  # HTTP requests for notifications
pytest>=7.0.0  # Testing framework
pytest-mock>=3.7.0  # Mocking for tests
plotly>=5.15.0
seaborn>=0.12.0
